Format cleanup

+++ simulation/halsim_ds_nt/src/main/native/cpp/HALSimDsNt.cpp
@@ -5,26 +5,31 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-#include <HALSimDsNt.h>
+#include "HALSimDsNt.h"
 
 void HALSimDSNT::Initialize() {
-  rootTable = nt::NetworkTableInstance::GetDefault().GetTable("sim")->GetSubTable("DS_CONTROL");  // Not to be confused with sim::DriverStation from HALSim LowFi
+  rootTable =
+      nt::NetworkTableInstance::GetDefault().GetTable("sim")->GetSubTable(
+          "DS_CONTROL");  // Not to be confused with sim::DriverStation from
+                          // HALSim LowFi
 
   // LOOP TIMING //
   auto timinghz = rootTable->GetEntry("timing_hz");
   timinghz.ForceSetDouble(50);
-  timinghz.AddListener([this](const nt::EntryNotification &ev) -> void {
-    double valIn = ev.value->GetDouble();
-    double val = 0;
-    val = (valIn < 1 ? 1 : valIn > 100 ? 100 : valIn);
+  timinghz.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        double valIn = ev.value->GetDouble();
+        double val = 0;
+        val = (valIn < 1 ? 1 : valIn > 100 ? 100 : valIn);
 
-    if (val != valIn) {
-      this->rootTable->GetEntry("timing_hz").ForceSetDouble(val);
-      Flush();
-    }
+        if (val != valIn) {
+          this->rootTable->GetEntry("timing_hz").ForceSetDouble(val);
+          Flush();
+        }
 
-    this->timingHz = val;
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+        this->timingHz = val;
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
   // MODES //
 
@@ -41,66 +46,80 @@ void HALSimDSNT::Initialize() {
   enabled.ForceSetBoolean(false);
   estop.ForceSetBoolean(false);
 
-  mtele.AddListener([this](const nt::EntryNotification &ev) -> void {
-    this->HandleModePress(HALSimDSNT_Mode::teleop, ev.value->GetBoolean());
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+  mtele.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        this->HandleModePress(HALSimDSNT_Mode::teleop, ev.value->GetBoolean());
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
-  mauto.AddListener([this](const nt::EntryNotification &ev) -> void {
-    this->HandleModePress(HALSimDSNT_Mode::auton, ev.value->GetBoolean());
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+  mauto.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        this->HandleModePress(HALSimDSNT_Mode::auton, ev.value->GetBoolean());
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
-  mtest.AddListener([this](const nt::EntryNotification &ev) -> void {
-    this->HandleModePress(HALSimDSNT_Mode::test, ev.value->GetBoolean());
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+  mtest.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        this->HandleModePress(HALSimDSNT_Mode::test, ev.value->GetBoolean());
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
-  enabled.AddListener([this](const nt::EntryNotification &ev) -> void {
-    std::lock_guard<wpi::mutex> lock(modeMutex);
-    if (!this->isEstop) {
-      this->isEnabled = ev.value->GetBoolean();
-    } else {
-      this->isEnabled = false;
-    }
-    this->UpdateModeButtons();
-    this->DoModeUpdate();
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+  enabled.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        std::lock_guard<wpi::mutex> lock(modeMutex);
+        if (!this->isEstop) {
+          this->isEnabled = ev.value->GetBoolean();
+        } else {
+          this->isEnabled = false;
+        }
+        this->UpdateModeButtons();
+        this->DoModeUpdate();
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
-  estop.AddListener([this](const nt::EntryNotification &ev) -> void {
-    std::lock_guard<wpi::mutex> lock(modeMutex);
-    this->isEstop = ev.value->GetBoolean();
-    if (this->isEstop) {
-      this->isEnabled = false;
-      this->UpdateModeButtons();
-    }
-    this->DoModeUpdate();
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+  estop.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        std::lock_guard<wpi::mutex> lock(modeMutex);
+        this->isEstop = ev.value->GetBoolean();
+        if (this->isEstop) {
+          this->isEnabled = false;
+          this->UpdateModeButtons();
+        }
+        this->DoModeUpdate();
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
   // ALLIANCES //
 
   allianceTable = rootTable->GetSubTable("alliance");
   auto allianceStation = allianceTable->GetEntry("station");
   auto allianceColorRed = allianceTable->GetEntry("red?");
-  
+
   allianceStation.ForceSetDouble(1);
   allianceColorRed.ForceSetBoolean(true);
 
-  allianceStation.AddListener([this](const nt::EntryNotification &ev) -> void {
-    double stnIn = ev.value->GetDouble();
-    double stn = 0;
-    stn = (stnIn > 3 ? 3 : stnIn < 1 ? 1 : stnIn);
+  allianceStation.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        double stnIn = ev.value->GetDouble();
+        double stn = 0;
+        stn = (stnIn > 3 ? 3 : stnIn < 1 ? 1 : stnIn);
 
-    if (stn != stnIn) {
-      this->allianceTable->GetEntry("station").ForceSetDouble(stn);
-      Flush();
-    }
+        if (stn != stnIn) {
+          this->allianceTable->GetEntry("station").ForceSetDouble(stn);
+          Flush();
+        }
 
-    this->allianceStation = stn;
-    this->DoAllianceUpdate();
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+        this->allianceStation = stn;
+        this->DoAllianceUpdate();
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
-  allianceColorRed.AddListener([this](const nt::EntryNotification &ev) -> void {
-    this->isAllianceRed = ev.value->GetBoolean();
-    this->DoAllianceUpdate();
-  }, NT_NotifyKind::NT_NOTIFY_UPDATE);
+  allianceColorRed.AddListener(
+      [this](const nt::EntryNotification& ev) -> void {
+        this->isAllianceRed = ev.value->GetBoolean();
+        this->DoAllianceUpdate();
+      },
+      NT_NotifyKind::NT_NOTIFY_UPDATE);
 
   // FINAL LOGIC //
 
@@ -127,9 +146,12 @@ void HALSimDSNT::HandleModePress(enum HALSimDSNT_Mode mode, bool isPressed) {
 }
 
 void HALSimDSNT::UpdateModeButtons() {
-  modeTable->GetEntry("teleop?").ForceSetBoolean(currentMode == HALSimDSNT_Mode::teleop);
-  modeTable->GetEntry("auto?").ForceSetBoolean(currentMode == HALSimDSNT_Mode::auton);
-  modeTable->GetEntry("test?").ForceSetBoolean(currentMode == HALSimDSNT_Mode::test);
+  modeTable->GetEntry("teleop?").ForceSetBoolean(currentMode ==
+                                                 HALSimDSNT_Mode::teleop);
+  modeTable->GetEntry("auto?").ForceSetBoolean(currentMode ==
+                                               HALSimDSNT_Mode::auton);
+  modeTable->GetEntry("test?").ForceSetBoolean(currentMode ==
+                                               HALSimDSNT_Mode::test);
   modeTable->GetEntry("enabled?").ForceSetBoolean(isEnabled);
   Flush();
 }
@@ -149,13 +171,17 @@ void HALSimDSNT::DoModeUpdate() {
 }
 
 void HALSimDSNT::DoAllianceUpdate() {
-  HALSIM_SetDriverStationAllianceStationId(static_cast<HAL_AllianceStationID>((isAllianceRed ? HAL_AllianceStationID_kRed1 : HAL_AllianceStationID_kBlue1) + (static_cast<int32_t>(allianceStation) - 1)));
+  HALSIM_SetDriverStationAllianceStationId(static_cast<HAL_AllianceStationID>(
+      (isAllianceRed ? HAL_AllianceStationID_kRed1
+                     : HAL_AllianceStationID_kBlue1) +
+      (static_cast<int32_t>(allianceStation) - 1)));
 }
 
 void HALSimDSNT::LoopFunc() {
   while (running) {
     double dt = 1000 / timingHz;
-    std::this_thread::sleep_for(std::chrono::milliseconds(static_cast<long long>(dt)));
+    std::this_thread::sleep_for(
+        std::chrono::milliseconds(static_cast<int64_t>(dt)));
     if (isEnabled) {
       currentMatchTime = currentMatchTime + dt;
       HALSIM_SetDriverStationMatchTime(currentMatchTime);
@@ -164,6 +190,4 @@ void HALSimDSNT::LoopFunc() {
   }
 }
 
-void HALSimDSNT::Flush() {
-  rootTable->GetInstance().Flush();
-}
+void HALSimDSNT::Flush() { rootTable->GetInstance().Flush(); }
